Add working happo integration
This commit will configure happo with the right things in order to get
the `happo run` command to succeed and create screenshots.

Getting the webpack config right turned out to be quite tricky. I
trial-and-errored my way forward by starting with a bare base config and
then slowly adding required properties. I think there might be a way to
get a default next config to pass in as `config` instead of having to
make up our own, but I'll leave that to a later time to figure out.

The tmpdir in .happo.js was necessary to avoid confusing build failures
where certain node_modules couldn't be found.